test permission case

+++ server/resource_methods_test.go
@@ -91,7 +91,7 @@ func GetResource() (*InMemoryTestDatabase, *resource.DbResource) {
 	resource.CheckErr(err, "Failed to update action table")
 
 	for _, table := range initConfig.Tables {
-		model := api2go.NewApi2GoModel(table.TableName, table.Columns, table.DefaultPermission, table.Relations)
+		model := api2go.NewApi2GoModel(table.TableName, table.Columns, int64(table.DefaultPermission), table.Relations)
 		res := resource.NewDbResource(model, wrapper, &ms, cruds, configStore, table)
 		cruds[table.TableName] = res
 	}
